% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_themes.R
\name{available_themes}
\alias{available_themes}
\title{Shows the registered themes.}
\usage{
available_themes(show = NULL)
}
\arguments{
\item{show}{Show only this specific.}
}
\description{
Shows the registered themes.
}
