% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corner_logo.R
\name{corner_logo}
\alias{corner_logo}
\title{Add corner logo}
\usage{
corner_logo(
  plot = ggplot2::last_plot(),
  logo = NULL,
  alpha = 1,
  position = "bottom-right",
  size = 1
)
}
\arguments{
\item{plot}{..}

\item{logo}{..}

\item{alpha}{..}

\item{position}{..}

\item{size}{..}
}
\value{
gg.
}
\description{
Add corner logo
}
\examples{

library(ggProfessional)

ggplot(iris, aes(Sepal.Length, Petal.Width)) +
geom_point() -
corner_logo

}
