% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_csv.R
\name{plot_to_csv}
\alias{plot_to_csv}
\title{Saves the plot to a new csv file (numbered).}
\usage{
plot_to_csv(plot = ggplot2::last_plot(), folder_name = "figures")
}
\arguments{
\item{plot}{..}

\item{folder_name}{..}
}
\description{
Saves the plot to a new csv file (numbered).
}
