% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_excel.R
\name{plot_to_excel}
\alias{plot_to_excel}
\title{Save figures to one excel file (numbered).}
\usage{
plot_to_excel(plot = ggplot2::last_plot(), file_name = "figures.xlsx")
}
\arguments{
\item{plot}{..}

\item{file_name}{..}
}
\description{
Save figures to one excel file (numbered).
}
