% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_to_ppt.R
\name{plot_to_ppt}
\alias{plot_to_ppt}
\title{Saves plots to one ppt file.}
\usage{
plot_to_ppt(
  plot = ggplot2::last_plot(),
  file_name = "figures.pptx",
  width = NULL,
  height = NULL
)
}
\arguments{
\item{plot}{..}

\item{file_name}{..}

\item{width}{..}

\item{height}{..}
}
\description{
Saves plots to one ppt file.
}
