% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_palette.R
\name{register_palette}
\alias{register_palette}
\title{Register a palette locally.}
\usage{
register_palette(pal, name)
}
\arguments{
\item{pal}{character vector (can be named).}

\item{name}{..}
}
\description{
Register a palette locally.
}
