% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_theme.R
\name{register_theme}
\alias{register_theme}
\title{Register a theme locally.}
\usage{
register_theme(theme, name)
}
\arguments{
\item{theme}{..}

\item{name}{..}
}
\description{
Register a theme locally.
}
