% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_palette.R
\name{set_palette}
\alias{set_palette}
\title{Set a palette as default for current sessoion.}
\usage{
set_palette(name, attach = FALSE, low = NULL, high = NULL)
}
\arguments{
\item{name}{..}

\item{attach}{Set as default?}
}
\description{
Set a palette as default for current sessoion.
}
