% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark_logo.R
\name{watermark_logo}
\alias{watermark_logo}
\title{Add watermark logo}
\usage{
watermark_logo(logo = NULL, alpha = 1)
}
\arguments{
\item{logo}{..}

\item{alpha}{..}

\item{plot}{..}

\item{position}{..}

\item{size}{..}
}
\value{
gg.
}
\description{
Add watermark logo
}
\examples{

library(ggProfessional)

ggplot(iris, aes(Sepal.Length, Petal.Width)) +
geom_point() +
watermark_logo()

}
